fixed math on stats book-keeping for overwrites (and duplicate inserts)
since I NEVER want to WALK the contents of the cache in order to determine
the stats, I have to be extremely careful with the accumulators for the various
resource and non-resource metrics, so every operation's change to the stats
is correctly accounted for.  this gets tricky in the overwrite and duplicate insert()
cases.  otherwise it's just what you expect- whne insert()ing something, add to the
accumulators.  when you delete(), subtract from the accumulators the per-entry related
stats.